Coordinate system improvments

+++ src/main/java/com/grough/automata/Box.java
@@ -4,9 +4,9 @@ package com.grough.automata;
  * Box with four corners.
  */
 public class Box {
-    int x1, y1, x2, y2, w, h;
+    float x1, y1, x2, y2, w, h;
 
-    Box(int x1, int y1, int x2, int y2) {
+    Box(float x1, float y1, float x2, float y2) {
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
